requirements.txt
numpy
pandas
matplotlib
pillow
streamlit
opencv-python
pytesseract
scikit-learn